--- complete_analysis.csv
@@ -1,1 +1,0 @@
-name,value,source,context,url,method,is_redirect_related,confidence,vulnerability_found
